[proliferate] organize automation worker surface
