+++ server/proliferate/server/automations/worker/__init__.py
@@ -1,0 +1,2 @@
+"""Automation worker process package."""
+
