+++ server/proliferate/server/automations/worker/main.py
@@ -4,14 +4,12 @@ from __future__ import annotations
 
 import argparse
 import asyncio
-import logging
 import signal
 from collections.abc import Sequence
 
 from proliferate.db import engine as db_engine
 from proliferate.db.migrations import validate_database_schema
 from proliferate.integrations.sentry import (
-    capture_server_sentry_exception,
     flush_server_sentry,
     init_server_sentry,
 )
@@ -20,12 +18,8 @@ from proliferate.server.automations.cloud_executor import (
     build_cloud_executor_config,
     run_cloud_executor_loop,
 )
-from proliferate.server.automations.service import run_scheduler_tick
+from proliferate.server.automations.worker.scheduler import run_scheduler_loop
 from proliferate.utils.logging import configure_server_logging
-
-logger = logging.getLogger(__name__)
-FAILURE_ESCALATION_THRESHOLD = 3
-MAX_FAILURE_BACKOFF_SECONDS = 300.0
 
 
 async def _validate_schema() -> None:
@@ -39,59 +33,6 @@ async def _validate_schema() -> None:
             "Could not connect to PostgreSQL for the automation worker. "
             "Start Postgres and run `make server-migrate` before starting the worker."
         ) from exc
-
-
-async def run_scheduler_loop(
-    *,
-    interval_seconds: float,
-    batch_size: int,
-    stop_event: asyncio.Event,
-) -> None:
-    logger.info(
-        "Automation scheduler worker started interval_seconds=%s batch_size=%s",
-        interval_seconds,
-        batch_size,
-    )
-    schema_validated = False
-    consecutive_failures = 0
-    while not stop_event.is_set():
-        try:
-            if not schema_validated:
-                await _validate_schema()
-                schema_validated = True
-            result = await run_scheduler_tick(batch_size=batch_size)
-            consecutive_failures = 0
-            if result.created_runs:
-                logger.info("Automation scheduler created runs count=%s", result.created_runs)
-            if result.swept_dispatching_runs:
-                logger.warning(
-                    "Automation scheduler swept dispatching runs count=%s",
-                    result.swept_dispatching_runs,
-                )
-            next_delay = interval_seconds
-        except Exception as exc:
-            consecutive_failures += 1
-            next_delay = min(
-                interval_seconds * (2 ** (consecutive_failures - 1)),
-                MAX_FAILURE_BACKOFF_SECONDS,
-            )
-            logger.exception(
-                "Automation scheduler tick failed consecutive_failures=%s next_delay_seconds=%s",
-                consecutive_failures,
-                next_delay,
-            )
-            if consecutive_failures >= FAILURE_ESCALATION_THRESHOLD:
-                capture_server_sentry_exception(
-                    exc,
-                    level="error",
-                    tags={"worker": "automation_scheduler"},
-                    extras={"consecutive_failures": consecutive_failures},
-                    fingerprint=["automation-scheduler", "tick-failed"],
-                )
-        try:
-            await asyncio.wait_for(stop_event.wait(), timeout=next_delay)
-        except TimeoutError:
-            continue
 
 
 def _install_signal_handlers(stop_event: asyncio.Event) -> None:
@@ -111,6 +52,7 @@ async def _amain(args: argparse.Namespace) -> None:
                 interval_seconds=args.interval_seconds,
                 batch_size=args.batch_size,
                 stop_event=stop_event,
+                validate_schema=_validate_schema,
             )
         elif args.role == "cloud-executor":
             await _validate_schema()
@@ -127,6 +69,7 @@ async def _amain(args: argparse.Namespace) -> None:
                         interval_seconds=args.interval_seconds,
                         batch_size=args.batch_size,
                         stop_event=stop_event,
+                        validate_schema=_validate_schema,
                     )
                 ),
                 asyncio.create_task(
